requirements.txt
asgiref==3.8.1
Django==5.1.3
django-crispy-forms==2.3
pillow==11.0.0
psycopg2==2.9.10
sqlparse==0.5.2
